stream: rename directfile.ts into stream_directfile.ts
