+++ src/features/list/directfile.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import directfile from "../../core/stream/directfile";
+import directfile from "../../core/stream/stream_directfile";
 import { IFeaturesObject } from "../types";
 
 /**
