+++ webpack-tests.config.js
@@ -56,7 +56,7 @@ const config = {
         HTML_SRT: JSON.stringify("../parsers/texttracks/srt/html.ts"),
         HTML_TTML: JSON.stringify("../parsers/texttracks/ttml/html/index.ts"),
         HTML_SAMI: JSON.stringify("../parsers/texttracks/sami/html.ts"),
-        DIRECTFILE: JSON.stringify("../core/stream/directfile.ts"),
+        DIRECTFILE: JSON.stringify("../core/stream/stream_directfile.ts"),
       },
       __DEV__: true,
       __LOGGER_LEVEL__: "\"NONE\"",
